Documentation to all headers

+++ include/impl/tensor/cpu/tensor.hpp
@@ -6,6 +6,7 @@
 #include <iostream>
 #include "impl/tensor/allocators.hpp"
 #include <omp.h>
+
 namespace lahva
 {
 
@@ -15,24 +16,28 @@ namespace lahva
     class Tensor
     {
     public:
-        /// @brief return size of Tensor
-        /// @return size of values
+        /// @brief Get number of elements in tensor
+        /// @return tensor size
         virtual size_t size() = 0;
-        /// @brief return pointer to the underlying data
-        /// @return ptr to begin of buffer
+
+        /// @brief Get pointer to tensor data
+        /// @return non-const pointer to tensor data
         virtual T *data() = 0;
 
-        /// @brief return size of Tensor
-        /// @return size of values
+        /// @brief Get number of elements in tensor
+        /// @return tensor size
         virtual size_t size() const = 0;
-        /// @brief return pointer to the underlying data
-        /// @return ptr to begin of buffer
+
+        /// @brief Get pointer to tensor data
+        /// @return const pointer to tensor data
         virtual T *data() const = 0;
 
+        /// @brief Compute sum of all tensor elements using parallel reduction (const version)
+        /// @return sum of all elements
         T sum() const
         {
             T res = (T)0;
-            #pragma omp parallel shared(res)
+#pragma omp parallel shared(res)
             {
                 T my_part = (T)0.0;
         
@@ -42,7 +47,7 @@ namespace lahva
                     my_part += this->data()[i];
                 }
         
-                #pragma omp critical
+#pragma omp critical
                 {
                     res += my_part;
                 }
@@ -51,55 +56,84 @@ namespace lahva
             return res;
         }
 
+        /// @brief Compute sum of all tensor elements using parallel reduction
+        /// @return sum of all elements
         T sum()
         {
             T res = (T)0;
-            #pragma omp parallel shared(res)
+#pragma omp parallel shared(res)
             {
                 T my_part = (T)0.0;
-        
+
                 #pragma omp for
-                for (long i = 0; i < this->size(); i++)
+                for (long i = 0; i < static_cast<long>(this->size()); i++)
                 {
                     my_part += this->data()[i];
                 }
-        
-                #pragma omp critical
+
+#pragma omp critical
                 {
                     res += my_part;
                 }
             }
             return res;
         }
+
+        /// @brief Const element access operator
+        /// @tparam D index type
+        /// @param[in] index element index
+        /// @return const reference to element at index
         template <typename D>
         const T &operator[](D index) const
         {
             return this->data()[static_cast<size_t>(index)];
         };
+
+        /// @brief Element access operator
+        /// @tparam D index type
+        /// @param[in] index element index
+        /// @return reference to element at index
         template <typename D>
         T &operator[](D index)
         {
             return this->data()[static_cast<size_t>(index)];
         };
 
-       // virtual const CPUAllocator<T> get_allocator() const = 0;
-
     };
 
+    /// @brief CPU-based tensor with host memory management
+    /// Base class for CPU tensors providing automatic memory management.
+    ///
+    /// @tparam T data type for tensor elements
+    /// @tparam Allocator host (CPU) memory allocator type (default: StdAllocator)
     template <typename T, typename Allocator = StdAllocator<T>>
     class CPUTensor : virtual public Tensor<T>
     {
         using alloc_ptr = CPUAllocator<T>;
 
     protected:
+        /// @brief Number of elements in the tensor
         std::size_t count_ = 0;
+
+        /// @brief Memory allocator instance
         Allocator alloc_;
+
+        /// @brief Rebound allocator for memory management
         std::shared_ptr<alloc_ptr> rebind_;
+
+        /// @brief Pointer to tensor data
         T *data_ = nullptr;
+
+        /// @brief Flag indicating if this object owns the data
         bool is_owner_ = true;
+
+        /// @brief Flag indicating if memory should not be allocated
         bool no_alloc = false;
 
     public:
+        /// @brief Construct CPU tensor with specified element count
+        /// @param[in] count number of elements in the tensor
+        /// @param[in] alloc host (CPU) memory allocator
         CPUTensor(size_t count, const alloc_ptr &alloc = Allocator()) : count_{count}, alloc_{alloc}
         {
             if (count_ > 0)
@@ -113,11 +147,19 @@ namespace lahva
                 data_ = nullptr;
             }
         };
+
+        /// @brief Construct CPU tensor without allocating initial memory
+        /// @param[in] alloc host (CPU) memory allocator
         CPUTensor(const alloc_ptr &alloc) : alloc_{alloc} { is_owner_ = false; };
+
+        /// @brief Default constructor for CPU tensor
         CPUTensor() : count_{0}, data_{nullptr}, no_alloc{true}
         {
-        
+
         };
+
+        /// @brief Copy constructor for CPU tensor
+        /// @param[in] other source tensor to copy
         CPUTensor(const CPUTensor &other) :  no_alloc{other.no_alloc}
         {
             this->count_ = other.count_;
@@ -135,6 +177,8 @@ namespace lahva
             
         };
 
+        /// @brief Move constructor for CPU tensor
+        /// @param[in] other source tensor to move from
         CPUTensor(CPUTensor &&other) : count_{other.count_}, alloc_{other.alloc_}
         {
 
@@ -145,6 +189,7 @@ namespace lahva
             other.count_ = 0;
         };
 
+        /// @brief Destructor for CPU tensor, releases memory
         virtual ~CPUTensor()
         {
             if (is_owner_)
@@ -153,11 +198,25 @@ namespace lahva
             }
         }
 
+        /// @brief Get pointer to tensor data
+        /// @return non-const pointer to tensor data
         T *data() override { return data_; };
+
+        /// @brief Get pointer to tensor data
+        /// @return const pointer to tensor data
         T *data() const override { return data_; };
+
+        /// @brief Get number of elements in tensor
+        /// @return tensor size
         size_t size() const override { return count_; };
+
+        /// @brief Get number of elements in tensor
+        /// @return tensor size
         size_t size() override { return count_; };
 
+        /// @brief Copy assignment operator
+        /// @param[in] other source tensor
+        /// @return reference to this tensor
         CPUTensor<T, Allocator> &operator=(const CPUTensor<T, Allocator> &other)
         {
             if (this != &other)
@@ -173,6 +232,9 @@ namespace lahva
             return *this;
         };
 
+        /// @brief Move assignment operator
+        /// @param[in] other source tensor to move from
+        /// @return reference to this tensor
         CPUTensor<T, Allocator> &operator=(CPUTensor &&other)
         {
             if (this != &other)
@@ -191,6 +253,8 @@ namespace lahva
             return *this;
         };
 
+        /// @brief Get memory allocator
+        /// @return allocator instance by reference
         const Allocator get_allocator() const { return alloc_; };
     };
